Setup Suggestions to be a dialog fragment and fixed some styling
