+++ app/src/main/res/layout/fragment_new_convention.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
             xmlns:app="http://schemas.android.com/apk/res-auto"
-            android:layout_width="300dp"
-            android:layout_height="300dp"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
             app:layout_behavior="@string/appbar_scrolling_view_behavior">
     <android.support.constraint.ConstraintLayout android:layout_width="300dp"
-                                                 android:layout_height="300dp">
+                                                 android:layout_height="wrap_content">
 
 
         <EditText
