+++ app/src/main/res/layout/fragment_new_convention_year.xml
@@ -6,7 +6,7 @@
             app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
     <android.support.constraint.ConstraintLayout
-        android:layout_width="match_parent"
+        android:layout_width="300dp"
         android:layout_height="wrap_content"
         android:orientation="vertical">
