+++ app/src/main/res/values/strings.xml
@@ -23,4 +23,5 @@
     <string name="logout">Logout</string>
     <string name="feedback_and_suggestions">Feedback and Suggestions</string>
     <string name="event_name">Event Name</string>
+    <string name="submit">Submit</string>
 </resources>
